+++ include/cask/deferred/PureDeferred.hpp
@@ -21,7 +21,7 @@ public:
     void onComplete(std::function<void(Either<T,E>)> callback) override;
     void onSuccess(std::function<void(T)> callback) override;
     void onError(std::function<void(E)> callback) override;
-    void onCancel(std::function<void()> callback) override;
+    void onCancel(const std::function<void()>& callback) override;
     void cancel() override;
     T await() override;
 private:
@@ -53,7 +53,7 @@ void PureDeferred<T,E>::onError(std::function<void(E)>) {
 }
 
 template <class T, class E>
-void PureDeferred<T,E>::onCancel(std::function<void()>) {
+void PureDeferred<T,E>::onCancel(const std::function<void()>&) {
     return;
 }